Refactor: Extract main window initialization logic into separate module. Set the window (win) and random_params global variables in the global scope of the module. Reformat code to adhere to PIP-8.
Separates window creation so that the files with app logic can be more focused and shorter. Canvas and Slider frame are made attributes to the window and can be accessed with the window.canvas and window.scales_label (any element could potentially be assigned to the window and accessed by it as it is possible thing to do with Python objects).

Main benefits:
- Separation of concerns
- Initialize global state variable `win` in the global scope of the module (other possibility is just to declare it in the module's scope for transparency, e.g. assign None to it).
- State of the canvas and sliders (`scales_label`) stored as attributes to the Tkinter window object and can be accessed with the `window.canvas` and `window.scales_label`